requirements.txt
# Requirements

configargparse==1.0
imageio==2.6.1
imageio-ffmpeg==0.3.0
numpy==1.18.1
pillow==7.0.0
pyyaml==5.3
scikit-image==0.16.2
scipy==1.4.1
torch==1.4.0
torchvision==0.5.0
tqdm==4.42.1
dotdict==0.1
tensorboard==1.14.0
tensorflow-gpu==1.14.0
